matchers: moving network matchers to their own extension
