+++ source/extensions/matching/network/application_protocol/config.h
@@ -1,0 +1,32 @@
+#include "envoy/extensions/matching/common_inputs/network/v3/network_inputs.pb.h"
+#include "envoy/extensions/matching/common_inputs/network/v3/network_inputs.pb.validate.h"
+#include "envoy/matcher/matcher.h"
+#include "envoy/network/filter.h"
+#include "envoy/registry/registry.h"
+
+#include "source/common/network/matching/inputs.h"
+#include "source/common/network/utility.h"
+
+namespace Envoy {
+namespace Network {
+namespace Matching {
+
+class ApplicationProtocolInput : public Matcher::DataInput<MatchingData> {
+public:
+  Matcher::DataInputGetResult get(const MatchingData& data) const override;
+};
+
+class ApplicationProtocolInputFactory
+    : public BaseFactory<
+          ApplicationProtocolInput,
+          envoy::extensions::matching::common_inputs::network::v3::ApplicationProtocolInput,
+          MatchingData> {
+public:
+  ApplicationProtocolInputFactory() : BaseFactory("application_protocol") {}
+};
+
+DECLARE_FACTORY(ApplicationProtocolInputFactory);
+
+} // namespace Matching
+} // namespace Network
+} // namespace Envoy
