+++ test/extensions/matching/network/common/BUILD
@@ -8,6 +8,7 @@ licenses(["notice"])  # Apache 2
 
 envoy_package()
 
+# For legacy reasons, the various network input tests are in one place.
 envoy_cc_test(
     name = "inputs_test",
     srcs = ["inputs_test.cc"],
@@ -17,6 +18,7 @@ envoy_cc_test(
         "//source/common/network/matching:data_impl_lib",
         "//source/common/network/matching:inputs_lib",
         "//source/common/router:string_accessor_lib",
+        "//source/extensions/matching/network/application_protocol:config",
         "//test/mocks/network:network_mocks",
     ],
 )
@@ -29,6 +31,7 @@ envoy_cc_test(
         "//source/common/network/matching:data_impl_lib",
         "//source/common/network/matching:inputs_lib",
         "//source/common/router:string_accessor_lib",
+        "//source/extensions/matching/network/application_protocol:config",
         "//test/common/matcher:test_utility_lib",
         "//test/mocks/matcher:matcher_mocks",
         "//test/mocks/network:network_mocks",
